+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2024-09-10
+### Added
+- PMC 2026 Announcement
+- Badges module
+- Early Bird Gold Membership link
+
+### Changed
+- Test Data Notices on Panels/Photo Ops page
+- Modules wording on main page
+- Minor visual tweaks
 
 ## [0.2.0e] - 2024-09-09
 ### Changed